iris example and main class

+++ run.sh
@@ -1,9 +1,29 @@
 #!/bin/bash
 
+if [[ $1 == "clean" ]]; then
+  $HADOOP_HOME/bin/hadoop fs -rm -r /TP
+  $HADOOP_HOME/bin/hadoop fs -mkdir /TP
+  $HADOOP_HOME/bin/hadoop fs -rm -r /TP/output
+  exit 0
+fi
+
+if [[ $1 == "add" ]]; then
+
+  $HADOOP_HOME/bin/hadoop fs -put data/institutions/child_care_centers.csv /TP
+  $HADOOP_HOME/bin/hadoop fs -put data/institutions/hospitals.csv /TP
+  $HADOOP_HOME/bin/hadoop fs -put data/institutions/local_law_enforcement_locations.csv /TP
+  $HADOOP_HOME/bin/hadoop fs -put data/institutions/places_of_worship.csv /TP
+  $HADOOP_HOME/bin/hadoop fs -put data/institutions/private_schools.csv /TP
+  $HADOOP_HOME/bin/hadoop fs -put data/institutions/public_schools.csv /TP
+  $HADOOP_HOME/bin/hadoop fs -put data/institutions/public_schools.csv /TP
+  $HADOOP_HOME/bin/hadoop fs -put data/gisjoin_data/cleaned_meta_data.csv /TP
+  exit 0
+fi
+
 gradle build
-$HADOOP_HOME/bin/hadoop fs -rm -r /TP
-$HADOOP_HOME/bin/hadoop fs -mkdir /TP
-$HADOOP_HOME/bin/hadoop fs -put testdata.txt /TP
+
+$HADOOP_HOME/bin/hadoop fs -put data/crime_data.csv /TP
+$HADOOP_HOME/bin/hadoop fs -put data/county_population.csv /TP
 $HADOOP_HOME/bin/hadoop fs -rm -r /TP/output
-$SPARK_HOME/bin/spark-submit --class cs455.TP.WordCount build/libs/TP-0.1.0.jar
+$SPARK_HOME/bin/spark-submit --class cs455.TP.Main build/libs/TP-0.1.0.jar
 $HADOOP_HOME/bin/hadoop fs -cat /TP/output/part-00002
